refactor: move root register code to apple(bonjour) module
