+++ src/os/apple.rs
@@ -1,1 +1,2 @@
+mod register;
 mod txt;
